PB-1142: destructure attempt-change-email module and add it plane on user module, and relate user by uuid
