+++ src/modules/user/exception/attempt-change-email-already-verified.exception.ts
