+++ src/modules/user/exception/attempt-change-email-has-expired.exception.ts
