+++ src/modules/user/exception/attempt-change-email-not-found.exception.ts
