+++ src/modules/user/exception/attempt-change-email.exception.ts
