+++ src/modules/user/user.usecase.ts
@@ -47,7 +47,10 @@ import { aes } from '@internxt/lib';
 import { PreCreatedUserAttributes } from './pre-created-users.attributes';
 import { PreCreatedUser } from './pre-created-user.domain';
 import { SequelizeSharingRepository } from '../sharing/sharing.repository';
-import { Transaction } from 'sequelize';
+import { AttemptChangeEmailRepository } from './attempt-change-email.repository';
+import { AttemptChangeEmailAlreadyVerifiedException } from './exception/attempt-change-email-already-verified.exception';
+import { AttemptChangeEmailHasExpiredException } from './exception/attempt-change-email-has-expired.exception';
+import { AttemptChangeEmailNotFoundException } from './exception/attempt-change-email-not-found.exception';
 
 class ReferralsNotAvailableError extends Error {
   constructor() {
@@ -114,6 +117,8 @@ export class UserUseCases {
     private readonly newsletterService: NewsletterService,
     private readonly keyServerRepository: SequelizeKeyServerRepository,
     private readonly avatarService: AvatarService,
+    private readonly attemptChangeEmailRepository: AttemptChangeEmailRepository,
+    private readonly mailerService: MailerService,
   ) {}
 
   findByEmail(email: User['email']): Promise<User | null> {
@@ -805,12 +810,8 @@ export class UserUseCases {
     return this.avatarService.getDownloadUrl(avatarKey);
   }
 
-  async changeUserEmailById(
-    userId: number,
-    newEmail: string,
-    transaction: Transaction,
-  ) {
-    const user = await this.userRepository.findById(userId);
+  async changeUserEmailById(userUuid: string, newEmail: string) {
+    const user = await this.userRepository.findByUuid(userUuid);
 
     if (!user) {
       throw new UserNotFoundError();
@@ -818,25 +819,89 @@ export class UserUseCases {
 
     const { uuid, email } = user;
 
-    await this.networkService.updateUserEmail(uuid, newEmail);
-    await this.userRepository.updateById(
-      user.id,
-      {
-        email: newEmail,
-        username: newEmail,
-        bridgeUser: newEmail,
-      },
-      transaction,
-    );
-    await this.sharedWorkspaceRepository.updateGuestEmail(
-      email,
-      newEmail,
-      transaction,
-    );
+    await this.userRepository.updateById(user.id, {
+      email: newEmail,
+      username: newEmail,
+      bridgeUser: newEmail,
+    });
+    await this.sharedWorkspaceRepository.updateGuestEmail(email, newEmail);
+
+    this.networkService.updateUserEmail(uuid, newEmail).catch(async () => {
+      await this.userRepository.updateById(user.id, {
+        email,
+        username: email,
+        bridgeUser: email,
+      });
+      await this.sharedWorkspaceRepository.updateGuestEmail(newEmail, email);
+    });
 
     return {
       oldEmail: user.email,
       newEmail: newEmail,
     };
+  }
+
+  async createAttemptChangeEmail(user: User, newEmail: string): Promise<void> {
+    const attempt = await this.attemptChangeEmailRepository.create({
+      userUuid: user.uuid,
+      newEmail,
+    });
+
+    const encryptedId = this.cryptoService.encryptText(attempt.id.toString());
+
+    await this.mailerService.sendUpdateUserEmailVerification(
+      newEmail,
+      encryptedId,
+    );
+  }
+
+  async isAttemptChangeEmailExpired(encryptedId: string) {
+    const attemptChangeEmailId = parseInt(
+      this.cryptoService.decryptText(encryptedId),
+    );
+
+    const attemptChangeEmail =
+      await this.attemptChangeEmailRepository.getOneById(attemptChangeEmailId);
+
+    if (!attemptChangeEmail) {
+      throw new AttemptChangeEmailNotFoundException();
+    }
+
+    if (attemptChangeEmail.isVerified) {
+      throw new AttemptChangeEmailAlreadyVerifiedException();
+    }
+
+    return { isExpired: attemptChangeEmail.isExpired };
+  }
+
+  async acceptAttemptChangeEmail(encryptedId: string) {
+    const attemptChangeEmailId = parseInt(
+      this.cryptoService.decryptText(encryptedId),
+    );
+
+    const attemptChangeEmail =
+      await this.attemptChangeEmailRepository.getOneById(attemptChangeEmailId);
+
+    if (!attemptChangeEmail) {
+      throw new AttemptChangeEmailNotFoundException();
+    }
+
+    if (attemptChangeEmail.isExpired) {
+      throw new AttemptChangeEmailHasExpiredException();
+    }
+
+    if (attemptChangeEmail.isVerified) {
+      throw new AttemptChangeEmailAlreadyVerifiedException();
+    }
+
+    const emails = await this.changeUserEmailById(
+      attemptChangeEmail.userUuid,
+      attemptChangeEmail.newEmail,
+    );
+    await this.attemptChangeEmailRepository.acceptAttemptChangeEmail(
+      attemptChangeEmailId,
+    );
+
+    return emails;
   }
 }
